Search Functionality and a (Somewhat) Cleaner Look
Added the ability to search through created recipes. At the moment,
double clicking on your search results will only print the recipe title
to the console.

Made the Recipe View and Search View somewhat cleaner, with the ability
to scroll through it if the application is too small. It still doesn't
stretch though.